Sat, Mar 16, 2024, 5:57 PM +05:30

+++ test_fuel/fuel.py
@@ -17,7 +17,7 @@ def convert(first):
             raise
         except ZeroDivisionError:
             raise
-    return int(x/y)*100
+    return (x/y)*100
 
 def gauge(p):
     if p<=1:
